+++ python/revng/internal/cli/_commands/model_migrate.py
@@ -12,9 +12,9 @@ import yaml
 
 from revng.internal.cli.commands_registry import Command, CommandsRegistry, Options
 from revng.internal.cli.revng import run_revng_command
-from revng.internal.cli.support import log_error
 from revng.model import YamlDumper  # type: ignore
 from revng.model.migrations import migrate
+from revng.support import log_error
 
 
 class ModelMigrateCommand(Command):
