+++ python/revng/internal/cli/support.py
@@ -9,7 +9,6 @@ import sys
 from contextlib import contextmanager
 from dataclasses import dataclass
 from io import BytesIO
-from shutil import which
 from subprocess import Popen
 from tarfile import open as tar_open
 from tempfile import NamedTemporaryFile
@@ -18,11 +17,10 @@ from typing import Tuple, Union
 
 import yaml
 
-from revng.internal.support import get_root, read_lines
-from revng.internal.support.collect import collect_files, collect_libraries
+from revng.internal.support.collect import collect_libraries
 from revng.internal.support.elf import is_executable
+from revng.support import get_command
 
-additional_bin_paths = read_lines(get_root() / "share/revng/additional-bin-paths")
 OptionalEnv = Optional[Mapping[str, str]]
 
 
@@ -39,10 +37,6 @@ class Options:
 
 def shlex_join(split_command: Iterable[str]) -> str:
     return " ".join(shlex.quote(arg) for arg in split_command)
-
-
-def log_error(msg: str):
-    sys.stderr.write(msg + "\n")
 
 
 def wrap(args: List[str], command_prefix: List[str]):
@@ -123,18 +117,6 @@ def exec_run(command, options: Options, environment: OptionalEnv) -> Union[int, 
         return 0
 
     return os.execvpe(command[0], command, environment)
-
-
-def get_command(command: str, search_prefixes: Iterable[str]) -> str:
-    for additional_bin_path in additional_bin_paths:
-        for executable in collect_files(search_prefixes, [additional_bin_path], command):
-            return executable
-
-    path = which(command)
-    if not path:
-        log_error(f'Couldn\'t find "{command}".')
-        assert False
-    return os.path.abspath(path)
 
 
 def interleave(base: List[str], repeat: str):
