+++ python/revng/internal/support/__init__.py
@@ -5,15 +5,13 @@
 import os
 from collections.abc import Iterable as CIterable
 from pathlib import Path
-from typing import Iterable, List, TypeVar, Union
+from typing import Iterable, TypeVar
 
 from xdg import xdg_cache_home
 
-T = TypeVar("T")
+from revng.support import SingleOrIterable
 
-SingleOrIterable = Union[T, Iterable[T]]
-AnyPath = Union[str, Path]
-AnyPaths = SingleOrIterable[AnyPath]
+T = TypeVar("T")
 
 
 def to_iterable(obj: SingleOrIterable[T]) -> Iterable[T]:
@@ -22,20 +20,8 @@ def to_iterable(obj: SingleOrIterable[T]) -> Iterable[T]:
     return (obj,)
 
 
-# We assume this file is in <root>/lib/python$VERSION/site-packages/revng/support
-def get_root() -> Path:
-    return (Path(__file__) / "../../../../../../../").resolve()
-
-
 def cache_directory() -> Path:
     if "REVNG_CACHE_DIR" in os.environ:
         return Path(os.environ["REVNG_CACHE_DIR"])
     else:
         return xdg_cache_home() / "revng"
-
-
-def read_lines(path: Path) -> List[str]:
-    if not path.exists():
-        return []
-    else:
-        return path.read_text().strip().split("\n")
